+++ apps/web-svelte/src/lib/server/media.ts
@@ -103,6 +103,9 @@ export async function fetchImages(
     tag?: string[];
     performer?: string[];
     studio?: string;
+    format?: string[];
+    animated?: string;
+    dimension?: string[];
     nsfw?: string;
     ratingMin?: number;
     ratingMax?: number;
@@ -122,6 +125,7 @@ export async function fetchImages(
       order: params?.order,
       gallery: params?.gallery,
       studio: params?.studio,
+      animated: params?.animated,
       nsfw: params?.nsfw,
       ratingMin: params?.ratingMin,
       ratingMax: params?.ratingMax,
@@ -135,6 +139,8 @@ export async function fetchImages(
     {
       tag: params?.tag,
       performer: params?.performer,
+      format: params?.format,
+      dimension: params?.dimension,
     },
   );